Add OwnableField class
  - Provides price(), owner(), changeOwner() and moveOn()
  - Abstract class, subclasses require rent()
