+++ server/source/serverlib/game/board/street.cpp
@@ -1,23 +1,12 @@
 #include "street.h"
 
-#include <game/game.h>
-
-#include <game/turn/buystate.h>
-#include <game/turn/paystate.h>
-
 Street::Street(const QString& name, int group, int price, int housePrice, std::array<int, ConstructionLevel::HOTEL + 1> rents)
-    : Field(name, FieldType::street)
+    : OwnableField(name, FieldType::street, price)
     , m_group(group)
-    , m_price(price)
     , m_housePrice(housePrice)
     , m_constructionLevel(ConstructionLevel::BASE)
     , m_rents(rents)
 {
-}
-
-void Street::changeOwner(const QString& newOwner)
-{
-    m_owner = newOwner;
 }
 
 void Street::changeConstructionLevel(const ConstructionLevel newLevel)
@@ -28,16 +17,6 @@ void Street::changeConstructionLevel(const ConstructionLevel newLevel)
 int Street::group() const
 {
     return m_group;
-}
-
-QString Street::owner() const
-{
-    return m_owner;
-}
-
-int Street::price() const
-{
-    return m_price;
 }
 
 int Street::housePrice() const
@@ -57,9 +36,8 @@ int Street::rent() const
 
 QJsonObject Street::description()
 {
-    auto description = Field::description();
+    auto description = OwnableField::description();
     description["group"] = m_group;
-    description["price"] = m_price;
     description["house_price"] = m_housePrice;
     description["rent"] = rents();
     return description;
@@ -74,23 +52,4 @@ QJsonArray Street::rents() const
     }
 
     return jsonRents;
-}
-
-bool Street::moveOn(const QString& playerName, Game* game)
-{
-    bool stateChanged = false;
-
-    if (m_owner == "")
-    {
-        game->stateChange<BuyState>();
-        stateChanged = true;
-    }
-    else if (m_owner != playerName)
-    {
-        std::vector<Debt> debts;
-        debts.emplace_back(playerName, m_owner, rent());
-        game->stateChange<PayState>(debts);
-        stateChanged = true;
-    }
-    return stateChanged;
 }
