+++ server/source/serverlib/game/board/street.h
@@ -6,7 +6,7 @@
 #include <QJsonObject>
 #include <QJsonArray>
 
-#include <game/board/field.h>
+#include <game/board/ownablefield.h>
 
 
 enum ConstructionLevel
@@ -19,21 +19,15 @@ enum ConstructionLevel
     HOTEL = 5,
 };
 
-class Street : public Field
+class Street : public OwnableField
 {
 public:
     Street(const QString& name, int group, int price, int housePrice, std::array<int, ConstructionLevel::HOTEL + 1> rents);
 
-    bool moveOn(const QString& playerName, Game* game) override;
-
     int group() const;
-    int price() const;
     int housePrice() const;
     ConstructionLevel constructionLevel() const;
-    int rent() const;
-    QString owner() const;
-
-	void changeOwner(const QString& newOwner);
+    int rent() const override;
     void changeConstructionLevel(const ConstructionLevel newLevel);
 
 	QJsonObject description() override;
@@ -41,9 +35,7 @@ public:
 protected:
     QJsonArray rents() const;
 
-    QString m_owner;
     const int m_group;
-    const int m_price;
     const int m_housePrice;
     ConstructionLevel m_constructionLevel;
     const std::array<int, ConstructionLevel::HOTEL + 1> m_rents;
